# api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Repository Metrics","description":"Repository analytics and metrics APIs"},{"name":"Pipeline Analysis","description":"AI-powered pipeline analysis APIs"},{"name":"Pipeline Runs","description":"Pipeline run management APIs"},{"name":"Webhook","description":"GitHub webhook receiver"},{"name":"GitHub","description":"GitHub repository browser APIs"},{"name":"Repositories","description":"Repository management APIs"},{"name":"Authentication","description":"GitHub OAuth + JWT Authentication APIs"}],"paths":{"/api/v1/settings/preferences":{"get":{"tags":["settings-controller"],"operationId":"getPrefs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["settings-controller"],"operationId":"savePrefs","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPreferences"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/settings/alerts":{"get":{"tags":["settings-controller"],"operationId":"alerts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["settings-controller"],"operationId":"saveAlerts","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPreferences"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/webhook/github":{"post":{"tags":["Webhook"],"operationId":"handleGitHub","parameters":[{"name":"X-GitHub-Event","in":"header","required":false,"schema":{"type":"string","default":"unknown"}},{"name":"X-Hub-Signature-256","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/settings/slack/test":{"post":{"tags":["settings-controller"],"operationId":"testSlack","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/settings/email/test":{"post":{"tags":["settings-controller"],"operationId":"testEmail","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/repositories":{"get":{"tags":["Repositories"],"operationId":"getRepositories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListRepositoryResponse"}}}}}},"post":{"tags":["Repositories"],"summary":"Add repository for monitoring","operationId":"addRepository","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddRepositoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRepositoryResponse"}}}}}}},"/api/v1/repositories/{repoId}/sync":{"post":{"tags":["Pipeline Runs"],"summary":"Trigger manual repository sync","operationId":"syncRepository","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/api/v1/repositories/{repoId}/runs/{runId}/analysis":{"get":{"tags":["Pipeline Analysis"],"summary":"Get AI analysis result","operationId":"getAnalysis","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"runId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RunAnalysis"}}}}}},"post":{"tags":["Pipeline Analysis"],"summary":"Trigger AI analysis for pipeline run","operationId":"triggerAnalysis","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"runId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/notifications/{id}/read":{"post":{"tags":["notification-controller"],"operationId":"markRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/notifications/read-all":{"post":{"tags":["notification-controller"],"operationId":"markAllRead","responses":{"200":{"description":"OK"}}}},"/api/v1/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh JWT access token","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout user session(s)","operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/v1/settings/integrations":{"get":{"tags":["settings-controller"],"operationId":"integrations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/repositories/{repositoryId}":{"get":{"tags":["Repositories"],"summary":"Get repository by id","operationId":"getRepository","parameters":[{"name":"repositoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRepositoryResponse"}}}}}},"delete":{"tags":["Repositories"],"summary":"Deactivate repository","operationId":"deactivateRepository","parameters":[{"name":"repositoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/repositories/{repoId}/runs":{"get":{"tags":["Pipeline Runs"],"summary":"Get paginated pipeline runs for a repository","operationId":"getRepositoryRuns","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0,"minimum":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20,"maximum":100,"minimum":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResponsePipelineRunResponse"}}}}}}},"/api/v1/repositories/{repoId}/runs/{runId}":{"get":{"tags":["Pipeline Runs"],"summary":"Get a single pipeline run by ID","operationId":"getPipelineRun","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"runId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PipelineRunResponse"}}}}}}},"/api/v1/repositories/{repoId}/metrics":{"get":{"tags":["Repository Metrics"],"summary":"Get repository metrics","operationId":"getRepositoryMetrics","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":14,"maximum":365,"minimum":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RepoMetricsDto"}}}}}}},"/api/v1/repositories/{repoId}/analyses":{"get":{"tags":["Pipeline Analysis"],"summary":"Get all repository analyses","operationId":"getRepositoryAnalyses","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RunAnalysis"}}}}}}}},"/api/v1/repositories/{repoId}/analyses/flaky":{"get":{"tags":["Pipeline Analysis"],"summary":"Get flaky pipeline analyses","operationId":"getFlakyAnalyses","parameters":[{"name":"repoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RunAnalysis"}}}}}}}},"/api/v1/notifications":{"get":{"tags":["notification-controller"],"operationId":"getNotifications","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}}},"/api/v1/notifications/unread-count":{"get":{"tags":["notification-controller"],"operationId":"unreadCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}},"/api/v1/github/repos":{"get":{"tags":["GitHub"],"summary":"Get authenticated user's GitHub repositories","operationId":"getRepositories_1","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1,"minimum":1}},{"name":"perPage","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":30,"maximum":100,"minimum":1}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"pushed"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListGitHubRepoDto"}}}}}}},"/api/v1/github/repos/search":{"get":{"tags":["GitHub"],"summary":"Search authenticated user's repositories","operationId":"searchRepositories","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListGitHubRepoDto"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current authenticated user","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/github/login":{"get":{"tags":["Authentication"],"summary":"Redirect user to GitHub OAuth","operationId":"login","responses":{"200":{"description":"OK"}}}},"/api/v1/auth/github/callback":{"get":{"tags":["Authentication"],"summary":"GitHub OAuth callback","operationId":"callback","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"MonitoredRepository":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"owner":{"type":"string"},"repoName":{"type":"string"},"source":{"type":"string","enum":["GITHUB_ACTIONS","JENKINS"]},"defaultBranch":{"type":"string"},"language":{"type":"string"},"accessToken":{"type":"string"},"webhookId":{"type":"integer","format":"int64"},"lastSyncedAt":{"type":"string","format":"date-time"},"totalRuns":{"type":"integer","format":"int32"},"successRate":{"type":"integer","format":"int32"},"lastRunStatus":{"type":"string"},"lastRunAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"},"github":{"type":"boolean"},"jenkins":{"type":"boolean"},"fullName":{"type":"string"}}},"User":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"githubId":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"avatarUrl":{"type":"string"},"githubUsername":{"type":"string"},"plan":{"type":"string"},"githubToken":{"type":"string"},"repositories":{"type":"array","items":{"$ref":"#/components/schemas/MonitoredRepository"}}}},"UserPreferences":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"darkMode":{"type":"boolean"},"emailNotifications":{"type":"boolean"},"slackNotifications":{"type":"boolean"},"aiInsights":{"type":"boolean"},"realtimeUpdates":{"type":"boolean"},"slackWebhookUrl":{"type":"string"},"flakyThreshold":{"type":"integer","format":"int32"},"failureThreshold":{"type":"integer","format":"int32"},"successRateThreshold":{"type":"integer","format":"int32"},"updatedAt":{"type":"string","format":"date-time"}}},"AddRepositoryRequest":{"type":"object","properties":{"githubRepoId":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"privateRepo":{"type":"boolean"},"htmlUrl":{"type":"string"},"owner":{"type":"string","maxLength":100,"minLength":0},"repoName":{"type":"string","maxLength":100,"minLength":0},"source":{"type":"string","enum":["GITHUB_ACTIONS","JENKINS"]},"defaultBranch":{"type":"string","maxLength":100,"minLength":0}},"required":["owner","repoName","source"]},"ApiResponseRepositoryResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/RepositoryResponse"},"error":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"type":"string"}},"timestamp":{"type":"string","format":"date-time"}}},"RepositoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"owner":{"type":"string"},"repoName":{"type":"string"},"source":{"type":"string","enum":["GITHUB_ACTIONS","JENKINS"]},"defaultBranch":{"type":"string"},"lastSyncedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"},"totalRuns":{"type":"integer","format":"int32"},"lastRunStatus":{"type":"string"},"lastRunAt":{"type":"string","format":"date-time"}}},"ApiResponseString":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"string"},"error":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"type":"string"}},"timestamp":{"type":"string","format":"date-time"}}},"RefreshTokenRequest":{"type":"object","properties":{"refresh_token":{"type":"string","maxLength":1024,"minLength":0}},"required":["refresh_token"]},"LogoutRequest":{"type":"object","properties":{"refresh_token":{"type":"string"},"logout_all":{"type":"boolean"},"session_id":{"type":"string"}}},"ApiResponseListRepositoryResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/RepositoryResponse"}},"error":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"type":"string"}},"timestamp":{"type":"string","format":"date-time"}}},"PagedResponsePipelineRunResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/PipelineRunResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}},"PipelineRunResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"externalRunId":{"type":"string"},"workflowName":{"type":"string"},"branch":{"type":"string"},"headSha":{"type":"string"},"status":{"type":"string","enum":["PENDING","RUNNING","SUCCESS","FAILED","QUEUED","CANCELLED","SKIPPED","UNKNOWN"]},"startedAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"},"durationMs":{"type":"integer","format":"int64"},"durationFormatted":{"type":"string"},"runUrl":{"type":"string"},"pullRequest":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"RunAnalysis":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"runId":{"type":"integer","format":"int64"},"repositoryId":{"type":"integer","format":"int64"},"stage":{"type":"string"},"failureCategory":{"type":"string","enum":["BUILD_COMPILATION","BUILD_DEPENDENCY","BUILD_TOOL","BUILD_TIMEOUT","TEST_UNIT","TEST_INTEGRATION","TEST_FLAKY","TEST_COVERAGE","DOCKER_BUILD","DOCKER_PUSH","DOCKER_PULL","CONTAINER_OOM","NETWORK_TIMEOUT","NETWORK_DNS","INFRA_RESOURCE","INFRA_RUNNER","K8S_DEPLOY","K8S_HEALTH_CHECK","K8S_IMAGE_PULL","K8S_CONFIG","LINTING","STATIC_ANALYSIS","SECURITY_SCAN","SECRET_DETECTION","AUTH_CREDENTIALS","AUTH_PERMISSIONS","CONFIG_MISSING_ENV","CONFIG_INVALID","ENV_MISMATCH","UNKNOWN"]},"severity":{"type":"string"},"priority":{"type":"string"},"summary":{"type":"string"},"rootCause":{"type":"string"},"diagnosis":{"type":"string"},"recommendation":{"type":"string"},"remediationSteps":{"type":"string"},"similarPatterns":{"type":"string"},"affectedComponent":{"type":"string"},"estimatedFixTime":{"type":"string"},"confidenceScore":{"type":"number","format":"double"},"classificationSource":{"type":"string"},"similarFailuresCount":{"type":"integer","format":"int32"},"flakinessScore":{"type":"number","format":"double"},"failingTests":{"type":"string"},"analysedAt":{"type":"string","format":"date-time"},"modelUsed":{"type":"string"},"flaky":{"type":"boolean"}}},"RepoMetricsDto":{"type":"object","properties":{"totalRuns":{"type":"integer","format":"int32"},"successfulRuns":{"type":"integer","format":"int32"},"failedRuns":{"type":"integer","format":"int32"},"runningRuns":{"type":"integer","format":"int32"},"pendingRuns":{"type":"integer","format":"int32"},"successRate":{"type":"integer","format":"int32"},"failureRate":{"type":"integer","format":"int32"},"avgDuration":{"type":"integer","format":"int32"},"failureRateTrend":{"type":"array","items":{"$ref":"#/components/schemas/TrendPoint"}},"stageDurations":{"type":"array","items":{"$ref":"#/components/schemas/StageDurationPoint"}},"statusBreakdown":{"type":"array","items":{"$ref":"#/components/schemas/StatusCount"}},"flakyTestCount":{"type":"integer","format":"int32"},"windowDays":{"type":"integer","format":"int32"}}},"StageDurationPoint":{"type":"object","properties":{"stage":{"type":"string"},"avgDuration":{"type":"integer","format":"int32"},"maxDuration":{"type":"integer","format":"int32"}}},"StatusCount":{"type":"object","properties":{"name":{"type":"string"},"value":{"type":"integer","format":"int32"}}},"TrendPoint":{"type":"object","properties":{"date":{"type":"string"},"failureRate":{"type":"integer","format":"int32"},"successRate":{"type":"integer","format":"int32"}}},"NotificationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"message":{"type":"string"},"type":{"type":"string"},"read":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListGitHubRepoDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/GitHubRepoDto"}},"error":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"type":"string"}},"timestamp":{"type":"string","format":"date-time"}}},"GitHubRepoDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"language":{"type":"string"},"full_name":{"type":"string"},"private":{"type":"boolean"},"default_branch":{"type":"string"},"owner_login":{"type":"string"},"html_url":{"type":"string"},"pushed_at":{"type":"string"},"stargazers_count":{"type":"integer","format":"int32"}}},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{},"error":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"type":"string"}},"timestamp":{"type":"string","format":"date-time"}}}}}}